Added Swagger UI, added validations to endpoints

+++ src/main/kotlin/management/UserService.kt
@@ -1,3 +1,7 @@
+package management
+
+import com.sun.research.ws.wadl.HTTPMethods
+import org.eclipse.jetty.http.HttpStatus
 import org.jdbi.v3.sqlobject.customizer.Bind
 import org.jdbi.v3.sqlobject.statement.SqlQuery
 import org.jdbi.v3.sqlobject.statement.SqlUpdate
@@ -19,7 +23,7 @@ interface UserDAO {
     fun findById(@Bind("id") id: Int): String?
 }
 
-object UserResource {
+object UserService { // service
     private var conn: Connection
 
     init {
@@ -27,12 +31,16 @@ object UserResource {
         conn = DriverManager.getConnection("jdbc:h2:~/test", "", "")
         val statement: Statement = conn.createStatement()
         statement.execute("drop table user if exists")
-        statement.execute("create table user(id int primary key, firstName varchar(100), lastName varchar(100), email varchar(100))")
-        statement.execute("insert into user values(1, 'UserName 1', 'UserSurname 1', 'email1@email.com')")
-        statement.execute("insert into user values(2, 'UserName 2', 'UserSurname 2', 'email2@email.com')")
-        statement.execute("insert into user values(3, 'UserName 3', 'UserSurname 3', 'email3@email.com')")
-        statement.execute("insert into user values(4, 'UserName 4', 'UserSurname 4', 'email4@email.com')")
-        val rs: ResultSet = statement.executeQuery("select * from user")
+        statement.execute("CREATE TABLE user(" +
+                              "id int NOT NULL AUTO_INCREMENT," +
+                              "firstName varchar(100)," +
+                              "lastName varchar(100)," +
+                              "email varchar(100)," +
+                              "PRIMARY KEY (id))")
+        statement.execute("insert into user (firstName, lastName, email) values('UserName 1', 'UserSurname 1', 'email1@email.com')")
+        statement.execute("insert into user (firstName, lastName, email) values('UserName 2', 'UserSurname 2', 'email2@email.com')")
+        statement.execute("insert into user (firstName, lastName, email) values('UserName 3', 'UserSurname 3', 'email3@email.com')")
+        statement.execute("insert into user (firstName, lastName, email) values('UserName 4', 'UserSurname 4', 'email4@email.com')")
         statement.close()
     }
 
@@ -44,9 +52,6 @@ object UserResource {
         if (rs.next()) {
             user = User(rs.getInt("id"), rs.getString("firstName"),
                         rs.getString("lastName"), rs.getString("email"))
-            println("User successfully retrieved")
-        } else {
-            println("No user with id = $id")
         }
         statement.close()
         return user
@@ -56,7 +61,7 @@ object UserResource {
         val result: MutableList<User?> = ArrayList<User?>()
         val statement: Statement = conn.createStatement()
         val rs: ResultSet = statement.executeQuery("select * from user")
-        var user: User? = null
+        var user: User
 
         while (rs.next()) {
             user = User(rs.getInt("id"), rs.getString("firstName"),
@@ -74,29 +79,58 @@ object UserResource {
         if (rs.next()) {
             count = rs.getInt(1)
         }
+        statement.close()
         return count
     }
 
-    fun remove(id: Int) {
-        val statement = conn.prepareStatement("delete from user where id = $id")
+    fun remove(id: Int): Int {
+        var result = HttpStatus.NOT_FOUND_404
+        var statement: Statement = conn.createStatement()
+        val rs: ResultSet = statement.executeQuery("select * from user where id = '$id'")
+        val notFound = !rs.next()
+        statement.close()
+
+        if (notFound) {
+            return result
+        }
+
+        statement = conn.prepareStatement("delete from user where id = $id")
         statement.executeUpdate()
         statement.close()
+        return HttpStatus.OK_200
     }
 
-    fun save(user: User): String {
-        var result = ""
+    fun update(user: User): String {
+        var result = HttpStatus.NOT_FOUND_404.toString()
         val statement: Statement = conn.createStatement()
-        val rs: ResultSet = statement.executeQuery("select * from user where id = ${user.id}")
+        val rs: ResultSet = statement.executeQuery("select * from user where id = '${user.id}'")
 
-        result = if (!rs.next()) {
-            statement.execute("insert into user values(${user.id}, '${user.firstName}', '${user.lastName}', '${user.email}')")
-            "Added User with id=" + user.id
-        } else {
-            statement.execute("update user " +
-                    "set id = ${user.id}, firstName = '${user.firstName}', lastName = '${user.lastName}', email = '${user.email}'" +
-                    "where id = ${user.id}"
-            )
-            "Updated User with id=" + user.id
+        if (rs.next()) {
+            val usersWithSameEmail: ResultSet = statement.executeQuery("select * from user where email = '${user.email}'")
+            result = if (usersWithSameEmail.next()) {
+                HttpStatus.CONFLICT_409.toString()
+            } else {
+                statement.execute(
+                    "update user " +
+                            "set firstName = '${user.firstName}', lastName = '${user.lastName}', email = '${user.email}'" +
+                            "where id = ${user.id}"
+                )
+                HttpStatus.OK_200.toString()
+            }
+        }
+        statement.close()
+
+        return result
+    }
+
+    fun create(user: UserCreation): String {
+        var result = HttpStatus.CONFLICT_409.toString()
+        val statement: Statement = conn.createStatement()
+        val rs: ResultSet = statement.executeQuery("select * from user where email = '${user.email}'")
+
+        if (!rs.next()) {
+            statement.execute("insert into user (firstName, lastName, email) values('${user.firstName}', '${user.lastName}', '${user.email}')")
+            result = HttpStatus.CREATED_201.toString()
         }
         statement.close()
 
@@ -104,9 +138,9 @@ object UserResource {
     }
 }
 
-//class UserResource(userDao: UserDAO, jdbi: Jdbi) {
+//class management.UserResource(userDao: management.UserDAO, jdbi: Jdbi) {
 //    private var conn: Connection
-//    private var userDao: UserDAO
+//    private var userDao: management.UserDAO
 //    private var jdbi: Jdbi
 //
 //    init {
@@ -127,15 +161,15 @@ object UserResource {
 //        statement.close()
 //    }
 //
-//    fun getById(id: Int): User? {
+//    fun getById(id: Int): management.User? {
 //        val statement: Statement = conn.createStatement()
 //        val rs: ResultSet = statement.executeQuery("select * from user where id = $id")
-//        var user: User? = null
+//        var user: management.User? = null
 //
 //        if (rs.next()) {
-//            user = User(rs.getInt("id"), rs.getString("firstName"),
+//            user = management.User(rs.getInt("id"), rs.getString("firstName"),
 //                        rs.getString("lastName"), rs.getString("email"))
-//            println("User successfully retrieved")
+//            println("management.User successfully retrieved")
 //        } else {
 //            println("No user with id = $id")
 //        }
@@ -144,13 +178,13 @@ object UserResource {
 //    }
 //
 //    fun all(): List<Any?> {
-//        val result: MutableList<User?> = ArrayList<User?>()
+//        val result: MutableList<management.User?> = ArrayList<management.User?>()
 //        val statement: Statement = conn.createStatement()
 //        val rs: ResultSet = statement.executeQuery("select * from user")
-//        var user: User? = null
+//        var user: management.User? = null
 //
 //        while (rs.next()) {
-//            user = User(rs.getInt("id"), rs.getString("firstName"),
+//            user = management.User(rs.getInt("id"), rs.getString("firstName"),
 //                        rs.getString("lastName"), rs.getString("email"))
 //            result.add(user)
 //        }
@@ -170,20 +204,20 @@ object UserResource {
 //        statement.close()
 //    }
 //
-//    fun save(user: User): String {
+//    fun save(user: management.User): String {
 //        var result = ""
 //        val statement: Statement = conn.createStatement()
 //        val rs: ResultSet = statement.executeQuery("select * from user where id = ${user.id}")
 //
 //        result = if (!rs.next()) {
 //            statement.execute("insert into user values(${user.id}, '${user.firstName}', '${user.lastName}', '${user.email}')")
-//            "Added User with id=" + user.id
+//            "Added management.User with id=" + user.id
 //        } else {
 //            statement.execute("update user " +
 //                    "set id = ${user.id}, firstName = '${user.firstName}', lastName = '${user.lastName}', email = '${user.email}')" +
 //                    "where id = ${user.id}"
 //            )
-//            "Updated User with id=" + user.id
+//            "Updated management.User with id=" + user.id
 //        }
 //        statement.close()
 //
